+++ src/FieldDefinition/Reference.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2020 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/factory-bot
+ */
+
+namespace Ergebnis\FactoryBot\FieldDefinition;
+
+use Ergebnis\FactoryBot\FixtureFactory;
+
+/**
+ * @internal
+ */
+final class Reference implements Resolvable
+{
+    private $className;
+
+    /**
+     * @param class-string $className
+     */
+    public function __construct(string $className)
+    {
+        $this->className = $className;
+    }
+
+    public function resolve(FixtureFactory $fixtureFactory): object
+    {
+        return $fixtureFactory->get($this->className);
+    }
+}