+++ test/Integration/FixtureFactoryTest.php
@@ -25,6 +25,7 @@ use Ergebnis\Test\Util\Helper;
  *
  * @uses \Ergebnis\FactoryBot\EntityDefinition
  * @uses \Ergebnis\FactoryBot\FieldDefinition
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Reference
  */
 final class FixtureFactoryTest extends AbstractTestCase
 {